# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ad Performance Tracker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/bigquery/performance": {
      "get": {
        "tags": [
          "bigquery"
        ],
        "summary": "Get Performance",
        "description": "Return deduplicated ad performance by employee acronym.\n\nBy default filters to P1 campaigns. When ``p1_only=false`` and dates are\nprovided, filters by the configured date column instead.",
        "operationId": "get_performance_api_bigquery_performance_get",
        "parameters": [
          {
            "name": "employee_acronym",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Acronym as __XX__ substring in ad_name (underscore-delimited)",
              "title": "Employee Acronym"
            },
            "description": "Acronym as __XX__ substring in ad_name (underscore-delimited)"
          },
          {
            "name": "p1_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter to P1 ads only. Set false for probationary date-range.",
              "default": true,
              "title": "P1 Only"
            },
            "description": "Filter to P1 ads only. Set false for probationary date-range."
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start of date range (YYYY-MM-DD). Used when p1_only=false.",
              "title": "Start Date"
            },
            "description": "Start of date range (YYYY-MM-DD). Used when p1_only=false."
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End of date range (YYYY-MM-DD). Used when p1_only=false.",
              "title": "End Date"
            },
            "description": "End of date range (YYYY-MM-DD). Used when p1_only=false."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Performance Api Bigquery Performance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bigquery/performance/summary": {
      "get": {
        "tags": [
          "bigquery"
        ],
        "summary": "Get Performance Summary",
        "description": "Return aggregated performance summary by employee acronym.\n\nBy default filters to P1 campaigns. When ``p1_only=false`` and dates are\nprovided, filters by the configured date column instead.",
        "operationId": "get_performance_summary_api_bigquery_performance_summary_get",
        "parameters": [
          {
            "name": "employee_acronym",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Acronym as __XX__ substring in ad_name (underscore-delimited)",
              "title": "Employee Acronym"
            },
            "description": "Acronym as __XX__ substring in ad_name (underscore-delimited)"
          },
          {
            "name": "p1_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter to P1 ads only. Set false for probationary date-range.",
              "default": true,
              "title": "P1 Only"
            },
            "description": "Filter to P1 ads only. Set false for probationary date-range."
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start of date range (YYYY-MM-DD). Used when p1_only=false.",
              "title": "Start Date"
            },
            "description": "Start of date range (YYYY-MM-DD). Used when p1_only=false."
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End of date range (YYYY-MM-DD). Used when p1_only=false.",
              "title": "End Date"
            },
            "description": "End of date range (YYYY-MM-DD). Used when p1_only=false."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Performance Summary Api Bigquery Performance Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}